test: compile the validator-ajv8 super-schema fixtures in-test
The precompiled-validator tests statically imported gitignored
superSchema*.cjs files, which forced generation to happen before vitest
loaded the tests — the reason the compileSchemas npm script, the vitest
globalSetup, and its pnpm/tsx subprocess (with its own --conditions flag)
all existed.

A compileSuperSchema() helper now compiles superSchema through the
file-writing compileSchemaValidators() API inside the test context (where
the @rjsf/source condition already applies), loads the generated module
via createRequire, and deletes the file. Coverage is unchanged: the
code-equality test still compares the string API against the file API's
output. The helper also single-sources the compile options from
testData's CUSTOM_OPTIONS, removing the keep-in-sync comment pair.
