--- packages/validator-ajv8/.gitignore
@@ -1,3 +1,0 @@
-# Ignore the generated test schema files
-test/harness/superSchema*.js
-test/harness/superSchema*.cjs
