+++ packages/validator-ajv8/test/createPrecompiledValidator.test.ts
@@ -1,12 +1,7 @@
-import type { RJSFSchema } from '@rjsf/utils';
-
-import type { Localizer, ValidatorFunctions } from '../src';
+import type { Localizer } from '../src';
 import createPrecompiledValidator from '../src/createPrecompiledValidator';
 import AJV8PrecompiledValidator from '../src/precompiledValidator';
-// oxlint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore file might be not yet generated by test/harness/compileTestSchema.js
-import * as superSchemaFns from './harness/superSchema.cjs'; // oxlint-disable-line import/extensions
-import superSchema from './harness/superSchema.json';
+import { compileSuperSchema, superSchema } from './harness/compileSuperSchema';
 
 vi.mock('../src/precompiledValidator');
 
@@ -15,8 +10,8 @@ interface TestType {
   bar: boolean;
 }
 
-const validateFns = superSchemaFns as unknown as ValidatorFunctions;
-const rootSchema = superSchema as unknown as RJSFSchema;
+const validateFns = compileSuperSchema();
+const rootSchema = superSchema;
 const mockedValidator = vi.mocked(AJV8PrecompiledValidator);
 
 describe('createPrecompiledValidator()', () => {
