+++ packages/validator-ajv8/test/precompiledValidator.test.ts
@@ -2,20 +2,14 @@ import type { ErrorSchema, FormValidation, RJSFSchema, RJSFValidationError, UiSc
 import { ErrorSchemaBuilder, JUNK_OPTION_ID, RJSF_REF_KEY, hashForSchema, noop, retrieveSchema } from '@rjsf/utils';
 import type { Mock } from 'vitest';
 
-import type { Localizer, ValidatorFunctions } from '../src';
+import type { Localizer } from '../src';
 import AJV8PrecompiledValidator from '../src/precompiledValidator';
-// oxlint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore file might be not yet generated by test/harness/compileTestSchema.cjs
-import * as superSchemaFns from './harness/superSchema.cjs'; // oxlint-disable-line import/extensions
-import superSchema from './harness/superSchema.json';
-// oxlint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore file might be not yet generated by test/harness/compileTestSchema.cjs
-import * as superSchemaOptionsFns from './harness/superSchemaOptions.cjs'; // oxlint-disable-line import/extensions
+import { SUPER_SCHEMA_OPTIONS, compileSuperSchema, superSchema } from './harness/compileSuperSchema';
 
-const validateFns = superSchemaFns as unknown as ValidatorFunctions;
+const validateFns = compileSuperSchema();
 
-const validateOptionsFns = superSchemaOptionsFns as unknown as ValidatorFunctions;
-const rootSchema = superSchema as unknown as RJSFSchema;
+const validateOptionsFns = compileSuperSchema(SUPER_SCHEMA_OPTIONS);
+const rootSchema = superSchema;
 
 describe('AJV8PrecompiledValidator', () => {
   let builder: ErrorSchemaBuilder;
