+++ packages/validator-ajv8/test/tsconfig.json
@@ -1,7 +1,6 @@
 {
   "extends": "../../../tsconfig.base.json",
   "include": ["./", "./**/*.json"],
-  "exclude": ["./harness/*.cjs"],
   "compilerOptions": {
     "rootDir": "./",
     "noEmit": true,
