+++ packages/validator-ajv8/vitest.config.ts
@@ -4,7 +4,6 @@ import base, { fullCoverage } from '../../testing/vitest.base';
 
 export default mergeConfig(base, {
   test: {
-    globalSetup: ['./test/harness/globalSetup.ts'],
     coverage: fullCoverage(),
   },
 });
